% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{add_validation_flags}
\alias{add_validation_flags}
\title{Add validation flags to catch data}
\usage{
add_validation_flags(catch_data)
}
\arguments{
\item{catch_data}{Processed catch data}
}
\value{
A dataframe with validation flags added
}
\description{
Adds quality control flags based on predefined thresholds and rules.
Flag descriptions:
1: Number of fishers is 0
2: Catch is negative
3: Catch is null despite group_catch being non-NULL
4: Catch is 0 despite group_catch being non-NULL
5: Number of buckets is too high (> 150)
6: Bucket is > 40kg
7: Small pelagic individual weight > 10kg
8: Individual weight > 250kg
9: Number of fishers is too high (>70) for non-ring nets
}
