% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable-helpers.R
\name{bulk_update_airtable}
\alias{bulk_update_airtable}
\title{Bulk Update Multiple Airtable Records}
\usage{
bulk_update_airtable(base_id, table_name, token, updates_df)
}
\arguments{
\item{base_id}{Character string. The Airtable base ID.}

\item{table_name}{Character string. Name of the table.}

\item{token}{Character string. Airtable API token.}

\item{updates_df}{Data frame with 'airtable_id' column and fields to update.}
}
\value{
List of response objects.
}
\description{
Updates multiple records in batches of 10.
}
