% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{cloud_storage_authenticate}
\alias{cloud_storage_authenticate}
\title{Authenticate to a Cloud Storage Provider}
\usage{
cloud_storage_authenticate(provider, options)
}
\arguments{
\item{provider}{A character string specifying the cloud provider ("gcs" or "aws").}

\item{options}{A named list of options specific to the cloud provider (see details).}
}
\description{
This function is primarily used internally by other functions to establish authentication
with specified cloud providers such as Google Cloud Services (GCS) or Amazon Web Services (AWS).
}
\details{
For GCS, the options list must include:
\itemize{
\item \code{service_account_key}: The contents of the authentication JSON file from your Google Project.
}

This function wraps \code{\link[googleCloudStorageR:gcs_auth]{googleCloudStorageR::gcs_auth()}} to handle GCS authentication.
}
\examples{
\dontrun{
authentication_details <- readLines("path/to/json_file.json")
cloud_storage_authenticate("gcs", list(service_account_key = authentication_details))
#'
}
}
\keyword{storage}
