% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable-helpers.R
\name{df_to_airtable}
\alias{df_to_airtable}
\title{Create New Airtable Records}
\usage{
df_to_airtable(df, base_id, table_name, token)
}
\arguments{
\item{df}{Data frame containing data to create.}

\item{base_id}{Character string. The Airtable base ID.}

\item{table_name}{Character string. Name of the table.}

\item{token}{Character string. Airtable API token.}
}
\value{
List of response objects.
}
\description{
Creates new records in batches of 10.
}
