% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{download_cloud_file}
\alias{download_cloud_file}
\title{Download Object from Cloud Storage}
\usage{
download_cloud_file(name, provider, options, file = name)
}
\arguments{
\item{name}{The name of the object in the storage bucket.}

\item{provider}{A character string specifying the cloud provider ("gcs" or "aws").}

\item{options}{A named list of provider-specific options including the bucket and authentication details.}

\item{file}{(Optional) The local path to save the downloaded object. If not specified, the object name is used.}
}
\value{
The path to the downloaded file.
}
\description{
Downloads an object from cloud storage to a local file.
}
\details{
For GCS, the options list should include:
\itemize{
\item \code{bucket}: The name of the bucket from which the object is downloaded.
\item \code{service_account_key}: The authentication JSON contents, if not previously authenticated.
}
}
\examples{
\dontrun{
authentication_details <- readLines("path/to/json_file.json")
download_cloud_file("object_name.json",
 "gcs",
 list(service_account_key = authentication_details, bucket = "my-bucket"),
 "local_path/to/save/object.json")
}

}
\keyword{storage}
