% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{download_parquet_from_cloud}
\alias{download_parquet_from_cloud}
\title{#' Download Parquet File from Cloud Storage}
\usage{
download_parquet_from_cloud(prefix, provider, options)
}
\arguments{
\item{prefix}{The file prefix path in cloud storage}

\item{provider}{The cloud storage provider key}

\item{options}{Cloud storage provider options}
}
\value{
A tibble containing the data from the parquet file
}
\description{
This function handles the process of downloading a parquet file from cloud storage
and reading it into memory.
}
\examples{
\dontrun{
raw_data <- download_parquet_from_cloud(
  prefix = conf$ingestion$koboform$catch$legacy$raw,
  provider = conf$storage$google$key,
  options = conf$storage$google$options
)
}

}
\keyword{storage}
