% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{extract_trip_ids_from_filenames}
\alias{extract_trip_ids_from_filenames}
\title{Extract Trip IDs from Track Filenames}
\usage{
extract_trip_ids_from_filenames(filenames)
}
\arguments{
\item{filenames}{Character vector of track filenames}
}
\value{
Character vector of trip IDs
}
\description{
Extract Trip IDs from Track Filenames
}
\keyword{internal}
