% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion-surveys.R
\name{flatten_field}
\alias{flatten_field}
\title{Flatten a Single Field of Kobotoolbox Data}
\usage{
flatten_field(x, p)
}
\arguments{
\item{x}{A vector or list representing a field in the data.}

\item{p}{The prefix or name associated with the field.}
}
\value{
Modified field, either unchanged, unnested, or appropriately renamed.
}
\description{
Processes each field within a row of survey data, handling both simple
vectors and nested lists.
}
\keyword{internal}
