% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{flatten_row}
\alias{flatten_row}
\title{Flatten a Single Row of Kobotoolbox Data}
\usage{
flatten_row(x)
}
\arguments{
\item{x}{A list representing a single row of Kobotoolbox data.}
}
\value{
A flattened tibble representing the input row.
}
\description{
Transforms each row of nested survey data into a flat tabular format.
}
\keyword{internal}
