% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-taxa.R
\name{get_fao_groups}
\alias{get_fao_groups}
\title{Extract and Format FAO Taxonomic Groups}
\usage{
get_fao_groups(fao_codes = NULL, asfis_list = NULL)
}
\arguments{
\item{fao_codes}{A character vector of FAO 3-alpha codes to extract. If NULL, returns empty dataset.}

\item{asfis_list}{A data frame containing the FAO ASFIS list with required columns:
\itemize{
\item Alpha3_Code - FAO 3-alpha code
\item Scientific_Name - Scientific name of taxon
\item English_name - Common name in English
\item Family - Family name
\item Order - Order name
\item ISSCAAP_Group - FAO ISSCAAP group number
}}
}
\value{
A tibble with standardized column names containing taxonomic information:
\itemize{
\item a3_code - FAO 3-alpha code
\item scientific_name - Scientific name
\item english_name - Common name in English
\item family - Family name
\item order - Order name
\item taxon_group - ISSCAAP group number
}
}
\description{
Filters and formats taxonomic information from the FAO ASFIS list for specified FAO 3-alpha codes,
excluding miscellaneous ("MZZ") and unknown ("UNKN") categories.
}
\details{
The function:
\enumerate{
\item Filters ASFIS list for specified FAO codes
\item Standardizes column names for consistency
\item Removes miscellaneous ("MZZ") and unknown ("UNKN") categories
\item Preserves the taxonomic hierarchy information
}
}
\note{
\itemize{
\item Requires dplyr package
\item MZZ (Miscellaneous marine fishes) and UNKN (Unknown) are automatically excluded
\item Column names are standardized for consistency with other functions
}
}
\examples{
# Example ASFIS data
asfis <- data.frame(
  Alpha3_Code = c("TUN", "MZZ", "RAG"),
  Scientific_Name = c("Thunnini", "Marine fishes nei", "Rastrelliger kanagurta"),
  English_name = c("Tunas", "Marine fishes", "Indian mackerel"),
  Family = c("SCOMBRIDAE", NA, "SCOMBRIDAE"),
  Order = c("PERCIFORMES", NA, "PERCIFORMES"),
  ISSCAAP_Group = c(36, 39, 37)
)

# Get taxonomic information for specific codes
fao_taxa <- get_fao_groups(c("TUN", "RAG"), asfis)

}
\keyword{mining}
