% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata tables}
\usage{
get_metadata(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The logging threshold level. Default is logger::DEBUG.}
}
\description{
Get Metadata tables from Google sheets. This function downloads
the tables include information about the fishery.
}
\details{
The parameters needed in \code{conf.yml} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
# Ensure you have the necessary configuration in conf.yml
metadata_tables <- get_metadata()
}
}
\keyword{storage}
