% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_preprocessed_surveys}
\alias{get_preprocessed_surveys}
\title{Download WCS preprocessed surveys}
\usage{
get_preprocessed_surveys(pars)
}
\arguments{
\item{pars}{The configuration file.}
}
\value{
A rds dataframe of preprocessed survey landings.
}
\description{
Download preprocessed WCS data from Google Cloud.
}
