% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{ingest_wcs_surveys}
\alias{ingest_wcs_surveys}
\title{Ingest WCS Catch Survey Data}
\usage{
ingest_wcs_surveys(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{Log level used as the threshold for logging (see \link[logger:log_levels]{logger::log_levels}).}
}
\value{
None; the function is used for its side effects.
}
\description{
This function automates the downloading of WCS fish catch survey data collected through Kobo Toolbox and uploads it to cloud storage services. The filenames are versioned to include date-time stamps and, if available, the first 7 digits of the Git commit SHA.
}
\details{
Configuration parameters required from \code{conf.yml} include:

\if{html}{\out{<div class="sourceCode">}}\preformatted{surveys:
  wcs_surveys:
    asset_id:
    username:
    password:
    file_prefix:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}

Progress through the function is tracked using the package \emph{logger}.

Logging progress is managed using the \code{logger} package.
}
\examples{
\dontrun{
ingest_wcs_surveys(logger::DEBUG)
}

}
\keyword{ingestion}
\keyword{workflow}
