% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-taxa.R
\name{match_species_from_taxa}
\alias{match_species_from_taxa}
\title{Match Species from Taxa Databases}
\usage{
match_species_from_taxa(species_list, taxa_data)
}
\arguments{
\item{species_list}{Processed species list from process_species_list()}

\item{taxa_data}{Taxa data from load_taxa_databases()}
}
\value{
A data frame with columns:
\itemize{
\item a3_code: FAO 3-alpha code
\item species: Scientific name
\item database: Source database
}
}
\description{
Matches species between FAO codes and database records, handling different taxonomic
levels (species, genus, family, order) appropriately.
}
\examples{
\dontrun{
taxa_data <- load_taxa_databases()
species_list <- process_species_list(fao_codes, asfis)
matches <- match_species_from_taxa(species_list, taxa_data)
}
}
\keyword{mining}
\keyword{preprocessing}
