% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{mdb_collection_push}
\alias{mdb_collection_push}
\title{Upload Data to MongoDB and Overwrite Existing Content}
\usage{
mdb_collection_push(
  data = NULL,
  connection_string = NULL,
  collection_name = NULL,
  db_name = NULL
)
}
\arguments{
\item{data}{A data frame containing the data to be uploaded.}

\item{connection_string}{A character string specifying the MongoDB connection URL.}

\item{collection_name}{A character string specifying the name of the collection.}

\item{db_name}{A character string specifying the name of the database.}
}
\value{
The number of data documents inserted into the collection (excluding the order document).
}
\description{
This function connects to a MongoDB database, removes all existing documents
from a specified collection, and then inserts new data. It also stores the
original column order to maintain data structure consistency.
}
\examples{
\dontrun{
# Upload and overwrite data in a MongoDB collection
result <- mdb_collection_push(
  data = processed_legacy_landings,
  connection_string = "mongodb://localhost:27017",
  collection_name = "my_collection",
  db_name = "my_database"
)
}

}
\keyword{storage}
