% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-wcs-surveys.R
\name{preprocess_wcs_surveys}
\alias{preprocess_wcs_surveys}
\title{Pre-process Zanzibar WCS surveys}
\usage{
preprocess_wcs_surveys(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See \link[logger:log_levels]{logger::log_levels} for more
details}
}
\value{
no outputs. This function is used for it's side effects
}
\description{
Downloads raw structured data from cloud storage services and pre-processes
into a binary format that is easier to deal with in R. During the pre-processing
phase, multiple columns in the survey data, which can become very wide due to
multiple recordings of similar information (e.g., species information), are nested
using a set of utility functions (\code{pt_nest_trip}, \code{pt_nest_catch},
\code{pt_nest_length}, \code{pt_nest_market}, \code{pt_nest_attachments}).
}
\details{
Nesting these columns helps in reducing the width of the dataframe and organizes
related columns into a single nested tibble column, thus simplifying subsequent
analysis and visualization tasks.#'

This function downloads the landings data from a given version (specified in
the config file \code{conf.yml}.The parameters needed are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{surveys:
  wcs_surveys:
    asset_id:
    username:
    password:
    file_prefix:
  version:
    preprocess:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}

Progress through the function is tracked using the package \emph{logger}.
}
\seealso{
\code{\link[=pt_nest_trip]{pt_nest_trip}}, \code{\link[=pt_nest_catch]{pt_nest_catch}},
\code{\link[=pt_nest_length]{pt_nest_length}}, \code{\link[=pt_nest_market]{pt_nest_market}},
\code{\link[=pt_nest_attachments]{pt_nest_attachments}}
}
\keyword{workflow}
