% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_wf_surveys}
\alias{preprocess_wf_surveys}
\title{Pre-process WorldFish Surveys}
\usage{
preprocess_wf_surveys(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The logging threshold to use. Default is logger::DEBUG.
See \code{logger::log_levels} for available options.}
}
\value{
None; the function is used for its side effects.
}
\description{
Downloads and preprocesses raw structured WorldFish survey data from cloud storage into a binary format.
The process includes organizing data from general and trip information into a structured format
optimized for analysis.
}
\details{
Configurations are read from \code{conf.yml} with the following necessary parameters:

\if{html}{\out{<div class="sourceCode">}}\preformatted{surveys:
  wf_surveys:
    asset_id:
    username:
    password:
    file_prefix:
  version:
    preprocess:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}

The function uses logging to track progress.
}
\keyword{preprocessing}
\keyword{workflow}
