% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{process_catch_data}
\alias{process_catch_data}
\title{Process catch data from surveys}
\usage{
process_catch_data(preprocessed_surveys, trips_info)
}
\arguments{
\item{preprocessed_surveys}{A preprocessed survey dataframe}

\item{trips_info}{Trip information dataframe from extract_trips_info}
}
\value{
A dataframe containing processed catch data
}
\description{
Processes and structures catch data from surveys, including catch numbers
and associated vessel/gear information.
}
