% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion-pds.R
\name{process_single_track}
\alias{process_single_track}
\title{Process Single PDS Track}
\usage{
process_single_track(trip_id, conf)
}
\arguments{
\item{trip_id}{Character. The ID of the trip to process.}

\item{conf}{List. Configuration parameters.}
}
\value{
List with processing status and details.
}
\description{
Process Single PDS Track
}
\keyword{internal}
