% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-taxa.R
\name{process_species_list}
\alias{process_species_list}
\title{Process Species List with Taxonomic Information}
\usage{
process_species_list(fao_codes, asfis_list)
}
\arguments{
\item{fao_codes}{Vector of FAO 3-alpha codes}

\item{asfis_list}{ASFIS list data frame containing taxonomic information}
}
\value{
A data frame with columns:
\itemize{
\item a3_code: FAO 3-alpha code
\item scientific_name: Scientific name (cleaned)
\item database: "fishbase" or "sealifebase"
\item rank: Taxonomic rank ("Genus", "Family", "Order", "Species")
\item ... (other taxonomic fields)
}
}
\description{
Processes a list of species by assigning database sources and taxonomic ranks.
Determines whether species should be looked up in FishBase or SeaLifeBase based
on their ISSCAAP group.
}
\note{
ISSCAAP groups 57, 45, 43, 42, 56 are assigned to SeaLifeBase;
all others to FishBase
}
\examples{
\dontrun{
species_list <- process_species_list(c("TUN", "PEZ"), asfis_data)
}
}
\keyword{mining}
\keyword{preprocessing}
