% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_nest_survey.R
\name{pt_nest_catch}
\alias{pt_nest_catch}
\title{Nest catch catch columns}
\usage{
pt_nest_catch(x)
}
\arguments{
\item{x}{Kobo survey in tabular format.}
}
\value{
Nested Landings data in which the information about multiple catch information has
been nested into a single column (\code{catch}) this column contains a
tibble for every row. This, attachment tibble has as many rows as there are
catch information.
}
\description{
Nests catch group columns obtained when reading structured data from the kobo
landings survey.
}
