% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcs-nesting.R
\name{pt_nest_length}
\alias{pt_nest_length}
\title{Nest Length Group Columns}
\usage{
pt_nest_length(x)
}
\arguments{
\item{x}{Data frame of WCS survey data in tabular format.}
}
\value{
A data frame with length data nested into a single 'length' column,
which contains a tibble for each row with multiple measurements.
}
\description{
Nests length group columns obtained from the structured data of WCS landings surveys.
This reduces the width of data by converting multiple related columns into a single nested column.
}
\keyword{internal}
