% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcs-nesting.R
\name{pt_nest_trip}
\alias{pt_nest_trip}
\title{Nest Trip Group Columns}
\usage{
pt_nest_trip(x)
}
\arguments{
\item{x}{A data frame containing structured survey data in tabular format.}
}
\value{
A data frame with trip data nested into a single 'trip' column containing a tibble for each row, corresponding to the various trip details.
}
\description{
Processes and nests trip-related columns from structured WCS landings survey data into a single 'trip' column. This approach consolidates trip information into nested tibbles within the dataframe, simplifying the structure for analysis.
}
\keyword{internal}
