% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{rename_child}
\alias{rename_child}
\title{Rename Nested Survey Data Elements}
\usage{
rename_child(x, i, p)
}
\arguments{
\item{x}{A list element, possibly nested, to be renamed.}

\item{i}{The index or key of the element within the parent list.}

\item{p}{The parent name to prepend.}
}
\value{
A renamed list element.
}
\description{
Appends a parent name or index to child elements within a nested list.
}
\keyword{internal}
