% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{upload_cloud_file}
\alias{upload_cloud_file}
\title{Upload File to Cloud Storage}
\usage{
upload_cloud_file(file, provider, options, name = file)
}
\arguments{
\item{file}{A character vector specifying the path(s) of the file(s) to upload.}

\item{provider}{A character string specifying the cloud provider ("gcs" or "aws").}

\item{options}{A named list of provider-specific options including the bucket and authentication details.}

\item{name}{(Optional) The name to assign to the file in the cloud. If not specified, the local file name is used.}
}
\value{
A list of metadata objects for the uploaded files if successful.
}
\description{
Uploads a local file to a specified cloud storage bucket, supporting both single and multiple files.
}
\details{
For GCS, the options list must include:
\itemize{
\item \code{bucket}: The name of the bucket to which files are uploaded.
\item \code{service_account_key}: The authentication JSON contents, if not previously authenticated.
}

This function utilizes \code{\link[googleCloudStorageR:gcs_upload]{googleCloudStorageR::gcs_upload()}} for file uploads to GCS.
}
\examples{
\dontrun{
authentication_details <- readLines("path/to/json_file.json")
upload_cloud_file(
  "path/to/local_file.csv",
  "gcs",
  list(service_account_key = authentication_details, bucket = "my-bucket")
)
}

}
\keyword{storage}
