% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{upload_parquet_to_cloud}
\alias{upload_parquet_to_cloud}
\title{Upload Processed Data to Cloud Storage}
\usage{
upload_parquet_to_cloud(
  data,
  prefix,
  provider,
  options,
  compression = "lz4",
  compression_level = 12
)
}
\arguments{
\item{data}{The data frame or tibble to upload}

\item{prefix}{The file prefix path in cloud storage}

\item{provider}{The cloud storage provider key}

\item{options}{Cloud storage provider options}

\item{compression}{Compression algorithm to use (default: "lz4")}

\item{compression_level}{Compression level (default: 12)}
}
\value{
Invisible NULL
}
\description{
This function handles the process of writing data to a parquet file and
uploading it to cloud storage.
}
\examples{
\dontrun{
upload_parquet_to_cloud(
  data = processed_data,
  prefix = conf$ingestion$koboform$catch$legacy$preprocessed,
  provider = conf$storage$google$key,
  options = conf$storage$google$options
)
}
}
\keyword{storage}
