% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_market}
\alias{validate_market}
\title{Validate Market Data and Detect Outliers in Price Per Kilogram}
\usage{
validate_market(data = NULL, k_max_price = NULL)
}
\arguments{
\item{data}{A data frame containing market information. Expected to include
columns \verb{_id}, \code{gear_type}, and \code{market}.}

\item{k_max_price}{Nonnegative constant used in the LocScaleB function,
based on the MAD method, to determine the
extension of bounds for outlier detection in price per kilogram. Common values
are 2, 2.5, and 3.}
}
\value{
A data frame with the original market data, additional columns for
detected outliers, and nested market data.
}
\description{
This function processes market data to identify outliers in the price per kilogram.
It selects necessary columns,
unnests market data, and calculates the price per kilogram. The data is grouped
by gear type, market group, and
species in the market. Outliers are detected using the \code{alert_outlier} function,
which applies the LocScaleB method
for outlier detection based on the Median Absolute Deviation (MAD) method.
The \code{k} parameter is used within the MAD
method to determine the extension of bounds for outlier detection.
}
\examples{
\dontrun{
# Assuming you have a data frame `market_data` with the necessary structure:
validated_market <- validate_market(data = market_data, k_max_price = 100)
}

}
