% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_wcs_surveys}
\alias{validate_wcs_surveys}
\title{Validate WCS Surveys Data}
\usage{
validate_wcs_surveys(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The logging level used as a threshold for the \code{logger} package, which controls the verbosity of logging output.}
}
\value{
None; the function is used for its side effects, which include data validation and uploading validated data to cloud storage.
}
\description{
Validates Wildlife Conservation Society (WCS) survey data by checking for inconsistencies in survey duration, catches, lengths, and market data. The function preprocesses surveys, performs validations, logs the process, and uploads the validated data to cloud storage.
}
\examples{
\dontrun{
validate_wcs_surveys(log_threshold = logger::INFO)
}
}
\seealso{
\code{\link{validate_catch}}, \code{\link{validate_length}}, \code{\link{validate_market}}
}
\keyword{validation}
\keyword{workflow}
