% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-taxa.R
\name{get_length_weight_batch}
\alias{get_length_weight_batch}
\title{Get Length-Weight Parameters for Species (Batch Version)}
\usage{
get_length_weight_batch(species_areas_filtered)
}
\arguments{
\item{species_areas_filtered}{Data frame with filtered species}
}
\value{
A data frame with columns:
\itemize{
\item a3_code: FAO 3-alpha code
\item species: Scientific name
\item area_code: FAO area code
\item database: Source database
\item type: Measurement type (e.g., "TL" for total length)
\item a: Length-weight parameter a
\item b: Length-weight parameter b
}
}
\description{
Retrieves length-weight relationship parameters for multiple species efficiently
by processing them in batches. Handles both fish and non-fish species appropriately.
}
\note{
\itemize{
\item For FishBase species, only total length (TL) measurements are used
\item Questionable estimates (EsQ = "yes") are excluded
}
}
\examples{
\dontrun{
lw_data <- get_length_weight_batch(species_areas_filtered)
}
}
\keyword{mining}
\keyword{preprocessing}
