% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata tables}
\usage{
get_metadata(table = NULL, log_threshold = logger::DEBUG)
}
\arguments{
\item{table}{Character. Name of the specific table to download. If NULL (default),
all tables specified in the configuration will be downloaded.}

\item{log_threshold}{The logging threshold level. Default is logger::DEBUG.}
}
\value{
A named list containing the requested tables as data frames. If a single
table is requested, the list will contain only that table. If no table is
specified, the list will contain all available tables.
}
\description{
Get Metadata tables from Google sheets. This function downloads
the tables that include information about the fishery. You can specify
a single table to download or get all available tables.
}
\details{
The parameters needed in \code{conf.yml} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
metadata:
  google_sheets:
    sheet_id:
    tables:
      - table1
      - table2
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
# Ensure you have the necessary configuration in conf.yml

# Download all metadata tables
metadata_tables <- get_metadata()

# Download a specific table
catch_table <- get_metadata(table = "devices")
}
}
\keyword{storage}
