% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_preprocessed_surveys}
\alias{get_preprocessed_surveys}
\title{Download Preprocessed Surveys}
\usage{
get_preprocessed_surveys(conf, prefix = NULL)
}
\arguments{
\item{conf}{A list representing the configuration settings, typically obtained from a YAML configuration file.}

\item{prefix}{A character string specifying the organization prefix to retrieve preprocessed surveys, either "wcs" or "ba".}
}
\value{
A dataframe of preprocessed survey landings, loaded from Parquet files.
}
\description{
Retrieves preprocessed survey data from Google Cloud Storage, specifically configured for WCS (Wildlife Conservation Society) datasets. This function fetches data stored in Parquet format.
}
\examples{
\dontrun{
config <- peskas.zanzibar.pipeline::read_config()
df_preprocessed <- get_preprocessed_surveys(config, prefix = "ba")
}

}
\keyword{storage}
