% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_validated_surveys}
\alias{get_validated_surveys}
\title{Download WCS Validated Surveys}
\usage{
get_validated_surveys(pars)
}
\arguments{
\item{pars}{A list representing the configuration settings, typically obtained from a YAML configuration file.}
}
\value{
A dataframe of validated survey landings, loaded from Parquet files.
}
\description{
Retrieves validated survey data from Google Cloud Storage, tailored for WCS (Wildlife Conservation Society) datasets. This function fetches data stored in Parquet format.
}
\examples{
\dontrun{
config <- peskas.zanzibar.pipeline::read_config()
df_validated <- get_validated_surveys(config)
#'
}

}
\keyword{storage}
