% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cloud-files.R
\name{get_validated_surveys}
\alias{get_validated_surveys}
\title{Download WCS validated surveys}
\usage{
get_validated_surveys(pars)
}
\arguments{
\item{pars}{The configuration file.}
}
\value{
A rds dataframe of validated survey landings.
}
\description{
Download validated WCS data from Google Cloud.
}
