% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-taxa.R
\name{load_taxa_databases}
\alias{load_taxa_databases}
\title{Load Taxa Data from FishBase and SeaLifeBase}
\usage{
load_taxa_databases()
}
\value{
A list with two elements:
\itemize{
\item fishbase: Data frame containing FishBase taxonomic data
\item sealifebase: Data frame containing SeaLifeBase taxonomic data
}
}
\description{
Retrieves taxonomic data from both FishBase and SeaLifeBase databases in a single function call.
This is typically the first step in species identification and classification.
}
\examples{
\dontrun{
taxa_data <- load_taxa_databases()
fishbase_taxa <- taxa_data$fishbase
sealifebase_taxa <- taxa_data$sealifebase
}
}
\keyword{mining}
\keyword{preprocessing}
