% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_wcs_surveys}
\alias{preprocess_wcs_surveys}
\title{Pre-process Zanzibar WCS Surveys}
\usage{
preprocess_wcs_surveys(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The logging threshold to use. Default is logger::DEBUG.
See \code{logger::log_levels} for available options.}
}
\value{
None; the function is used for its side effects.
}
\description{
Downloads and preprocesses raw structured WCS survey data from cloud storage into a binary format. The process includes nesting multiple columns related to species information into single columns within a dataframe, which helps reduce its width and organize data efficiently for analysis.
}
\details{
Configurations are read from \code{config.yml} with the following necessary parameters:

\if{html}{\out{<div class="sourceCode">}}\preformatted{surveys:
  wcs_surveys:
    asset_id:
    username:
    password:
    file_prefix:
  version:
    preprocess:
storage:
  storage_name:
    key:
    options:
      project:
      bucket:
      service_account_key:
}\if{html}{\out{</div>}}

The function uses logging to track progress.
}
\seealso{
\code{\link[=pt_nest_trip]{pt_nest_trip}}, \code{\link[=pt_nest_catch]{pt_nest_catch}},
\code{\link[=pt_nest_length]{pt_nest_length}}, \code{\link[=pt_nest_market]{pt_nest_market}},
\code{\link[=pt_nest_attachments]{pt_nest_attachments}}
}
\keyword{preprocessing}
\keyword{workflow}
