% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R, R/preprocessing_BACKUP.R
\name{process_version_data}
\alias{process_version_data}
\title{Process Version Data Helper Function}
\usage{
process_version_data(catch_info, general_info, asfis)

process_version_data(catch_info, general_info, asfis)
}
\arguments{
\item{catch_info}{Processed catch information}

\item{general_info}{Processed general information}

\item{asfis}{ASFIS species data}
}
\value{
Combined and processed survey data

Combined and processed survey data
}
\description{
Internal helper function to process catch and general info for a specific survey version

Internal helper function to process catch and general info for a specific survey version
}
\keyword{internal}
