% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pt_nest_catch}
\alias{pt_nest_catch}
\title{Nest Catch Group Columns}
\usage{
pt_nest_catch(x)
}
\arguments{
\item{x}{Data frame of WCS survey data in tabular format.}
}
\value{
A data frame with catch data nested into a 'catch' column, containing a tibble
for each row with various catch-related attributes.
}
\description{
Nests catch group columns from WCS structured survey data to organize multiple
related catch data points into a single nested 'catch' column per row.
}
\keyword{internal}
