% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_nest_survey.R
\name{pt_nest_length}
\alias{pt_nest_length}
\title{Nest length group columns}
\usage{
pt_nest_length(x)
}
\arguments{
\item{x}{Kobo survey in tabular format.}
}
\value{
Nested Landings data in which the information about multiple length information has
been nested into a single column (\code{length}) this column contains a
tibble for every row. This, attachment tibble has as many rows as there are
length information
}
\description{
Nests length group columns obtained when reading structured data from the kobo
landings survey.
}
