% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_nest_survey.R
\name{pt_nest_market}
\alias{pt_nest_market}
\title{Nest market group columns}
\usage{
pt_nest_market(x)
}
\arguments{
\item{x}{Kobo survey in tabular format.}
}
\value{
Nested Landings data in which the information about multiple market information has
been nested into a single column (\code{market}) this column contains a
tibble for every row. This, attachment tibble has as many rows as there are
market information.
}
\description{
Nests market group columns obtained when reading structured data from the kobo
landings survey.
}
