% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_nest_survey.R
\name{pt_nest_trip}
\alias{pt_nest_trip}
\title{Nest trip catch columns}
\usage{
pt_nest_trip(x)
}
\arguments{
\item{x}{Kobo survey in tabular format.}
}
\value{
Nested Landings data in which the information about multiple trip information has
been nested into a single column (\code{trip}) this column contains a
tibble for every row. This, attachment tibble has as many rows as there are
trip information.
}
\description{
Nests trip group columns obtained when reading structured data from the kobo
landings survey.
}
