% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{retrieve_wcs_surveys}
\alias{retrieve_wcs_surveys}
\title{Retrieve WCS Surveys from Kobotoolbox}
\usage{
retrieve_wcs_surveys(
  prefix = NULL,
  append_version = NULL,
  url = NULL,
  project_id = NULL,
  username = NULL,
  psswd = NULL,
  encoding = NULL
)
}
\arguments{
\item{prefix}{Filename prefix or path for downloaded files.}

\item{append_version}{Boolean indicating whether to append versioning info to filenames.}

\item{url}{URL of the Kobotoolbox instance.}

\item{project_id}{Project asset ID for data download.}

\item{username}{Kobotoolbox account username.}

\item{psswd}{Kobotoolbox account password.}

\item{encoding}{Character encoding for the downloaded data; defaults to "UTF-8".}
}
\value{
Vector of paths for the downloaded Parquet files.
}
\description{
Downloads survey data from Kobotoolbox for a specified project and uploads the data in Parquet format. File naming can include versioning details.
}
\examples{
\dontrun{
file_list <- retrieve_wcs_surveys(
  prefix = "my_data",
  append_version = TRUE,
  url = "kf.kobotoolbox.org",
  project_id = "my_project_id",
  username = "admin",
  psswd = "admin",
  encoding = "UTF-8"
)
}

}
\keyword{ingestion}
