% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_surveys_time}
\alias{validate_surveys_time}
\title{Validate surveys' temporal parameters}
\usage{
validate_surveys_time(data, hrs_max = NULL, hrs_min)
}
\arguments{
\item{data}{A preprocessed data frame}

\item{hrs_max}{Upper threshold of fishing trip duration.}

\item{hrs_min}{Lower threshold of fishing trip duration.}
}
\value{
A list containing data frames with validated catch dates and catch
duration.
}
\description{
This function takes a preprocessed landings' matrix and validate temporal
info associated to each survey.
}
\examples{
\dontrun{
pars <- read_config()
landings <- get_preprocessed_surveys(pars)
validate_surveys_time(landings, hrs_max = 72, hrs_min = 1)
}
}
