% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-surveys.R
\name{validate_wcs_surveys}
\alias{validate_wcs_surveys}
\title{Validate WCS Surveys Data}
\usage{
validate_wcs_surveys(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{A log level threshold from the \code{logger} package, used to
set the minimum level of log messages to be captured.}
}
\value{
No return value; this function is called for its side effects,
including data validation, file creation, and cloud uploading.
}
\description{
This function validates Wildlife Conservation Society (WCS) survey data. It reads
configuration parameters, preprocesses surveys,
and performs various validations on survey duration, catches, lengths, and market data.
It also logs the process at specified
log thresholds. The function consolidates validated data and saves it as an RDS file,
which is then uploaded to cloud storage.
}
\examples{
# Assuming necessary configuration and data are available:
\dontrun{
validate_wcs_surveys(log_threshold = logger::INFO)
}
}
\seealso{
\code{\link{validate_catch}}, \code{\link{validate_length}}, \code{\link{validate_market}}
}
